Moves Project to its own conventional file
